requirements.txt
numpy
opencv-python
PyQt5
easyocr
jaro-winkler
